Renamed examples/bov -> cmd/bcpd
