+++ cmd/bcpd/app/codec.proto
