+++ cmd/bcpd/app/examples.go
