+++ cmd/bcpd/commands/docs.go
