+++ cmd/bcpd/commands/testdata/genesis.json
