+++ cmd/bcpd/commands/testdata/priv_validator.json
